+++ lib/kitchen/directions/bake_first_elements.rb
@@ -2,10 +2,10 @@
 
 module Kitchen
   module Directions
-    module BakeProblemFirstElements
+    module BakeFirstElements
       def self.v1(within:, selectors:)
         selectors.each do |selector|
-          within.search(".os-problem-container > #{selector}:first-child").each do |problem|
+          within.search("#{selector}:first-child").each do |problem|
             problem.add_class('first-element')
             problem.parent.add_class('has-first-element')
           end