+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [3.2.0] - 2021-04-19
 
+* Refactors `BakeFirstElements` and `BakeNumberedExercise` (minor)
 * Adds a decorating option of equation numbers on `BakeEquations` (minor)
 
 ## [3.1.0] - 2021-04-19